Fixed the range error
The program had the error of calling the ranger multiple times and would
return -1 as an error. I fixed this by adding the wait() function before
each count check in the main function. I also added the wait() function
in the avoid_crash() function to stop this error as well. There is still
the error of the motor is not working. I used a different car and the
error was still there so it must be a software or wiring error.