Fixed the motor problem and added features
The motor problem has been fixed; the PCA counting module for the
motor's pulse was not init so it didn't work. The following features
were added: the RF-transceiver works,, the lcd display shows direction,
range, pw percentage, and battery level, and adjusted the motor boolean
statements to adjust speed.

The init of steering after the direction needs to be fixed and when it
starts to turn to the left sharply.