Successful Build
The code runs: the car moves forward going up a slope and backwards
going down a slope. The car turns right when it is lean down the slope
on the left and vice versa. The steering value is also added into the
driving because the goal is to drive to a flat surface and if it is
steering, then it is not on a flat surface.

We are now experimenting with gains to find the best values.